app/requirements.txt
Flask==3.0.3
mysqlclient==2.2.4
Flask-MySQLdb==2.0.0
selenium==4.24.0
python-dotenv==1.0.1
webdriver-manager==4.0.2